@@ -16,10 +16,10 @@
         }
     },
     "require": {
-        "php": ">=5.5.9"
+        "php": ">=8.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.0"
+        "phpunit/phpunit": "^9.0|^10.0"
     },
     "suggest": {
         "ext-pcntl": "*",
